feat: added get user query

+++ PERUSTARS/PERUSTARS/Shared/Domain/Repositories/IBaseRepository.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using System.Numerics;
 using System.Threading.Tasks;
 
 namespace PERUSTARS.Shared.Domain.Repositories
@@ -6,7 +7,7 @@ namespace PERUSTARS.Shared.Domain.Repositories
     public interface IBaseRepository<TEntity>
     {
         Task AddAsync(TEntity entity);
-        Task<TEntity?> FindByIdAsync(int id);
+        Task<TEntity?> FindByIdAsync(BigInteger id);
         void Update(TEntity entity);
         void Remove(TEntity entity);
         Task<IEnumerable<TEntity>> ListAsync();
